requirements.txt
matplotlib>=3.8.0
pillow>=10.0.0
moviepy>=1.0.3
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
elevenlabs>=1.0.0
openai>=1.0.0
rich>=13.0.0
